+++ openzwavemqtt/const.py
@@ -1,3 +1,7 @@
+import logging
+
+LOGGER = logging.getLogger("openzwavemqtt")
+
 EVENT_PLACEHOLDER = "missing"
 EVENT_INSTANCE_ADDED = "instance_added"
 EVENT_INSTANCE_CHANGED = "instance_changed"
@@ -13,3 +17,5 @@ EVENT_NODE_REMOVED = "node_removed"
 EVENT_VALUE_ADDED = "value_added"
 EVENT_VALUE_CHANGED = "value_changed"
 EVENT_VALUE_REMOVED = "value_removed"
+
+EMPTY = {}
